# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Course-scoped API for Study Buddy AI
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [courses]
      summary: Get the authenticated learner dashboard summary
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /courses:
    get:
      operationId: listCourses
      tags: [courses]
      summary: List the authenticated learner's courses
      responses:
        "200":
          description: Courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createCourse
      tags: [courses]
      summary: Create a course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseInput"
      responses:
        "201":
          description: Created course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /courses/{courseId}:
    parameters:
      - $ref: "#/components/parameters/CourseId"
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get one course
      responses:
        "200":
          description: Course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateCourse
      tags: [courses]
      summary: Update a course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseUpdate"
      responses:
        "200":
          description: Updated course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteCourse
      tags: [courses]
      summary: Delete a course and its scoped records
      responses:
        "204":
          description: Course deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /courses/{courseId}/materials:
    parameters:
      - $ref: "#/components/parameters/CourseId"
    get:
      operationId: listMaterials
      tags: [materials]
      summary: List materials scoped to a course
      responses:
        "200":
          description: Course materials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Material"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      operationId: createMaterial
      tags: [materials]
      summary: Register an uploaded course material
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaterialInput"
      responses:
        "201":
          description: Material processing record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /courses/{courseId}/materials/{materialId}:
    parameters:
      - $ref: "#/components/parameters/CourseId"
      - $ref: "#/components/parameters/MaterialId"
    delete:
      operationId: deleteMaterial
      tags: [materials]
      summary: Delete a course material
      responses:
        "204":
          description: Material deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /courses/{courseId}/sessions:
    parameters:
      - $ref: "#/components/parameters/CourseId"
    get:
      operationId: listStudySessions
      tags: [sessions]
      summary: List study sessions scoped to a course
      responses:
        "200":
          description: Study sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudySession"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      operationId: createStudySession
      tags: [sessions]
      summary: Schedule a study session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudySessionInput"
      responses:
        "201":
          description: Scheduled session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudySession"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /study-sessions/{sessionId}:
    parameters:
      - $ref: "#/components/parameters/SessionId"
    patch:
      operationId: updateStudySession
      tags: [sessions]
      summary: Edit a study session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudySessionUpdate"
      responses:
        "200":
          description: Updated session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudySession"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteStudySession
      tags: [sessions]
      summary: Delete a study session
      responses:
        "204":
          description: Session deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /courses/{courseId}/tutor:
    parameters:
      - $ref: "#/components/parameters/CourseId"
    post:
      operationId: askTutor
      tags: [tutor]
      summary: Ask the course-aware AI tutor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TutorQuestion"
      responses:
        "200":
          description: Tutor answer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TutorAnswer"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          description: AI service is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /courses/{courseId}/progress:
    parameters:
      - $ref: "#/components/parameters/CourseId"
    get:
      operationId: getCourseProgress
      tags: [learning]
      summary: Get adaptive progress for a course
      responses:
        "200":
          description: Course progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseProgress"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /courses/{courseId}/learning/next:
    parameters:
      - $ref: "#/components/parameters/CourseId"
    post:
      operationId: getNextLearningQuestion
      tags: [learning]
      summary: Generate the next scheduled active-learning question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LearningQuestionRequest"
      responses:
        "200":
          description: Next question or no question due
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearningQuestion"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          description: AI service is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /courses/{courseId}/learning/answers:
    parameters:
      - $ref: "#/components/parameters/CourseId"
    post:
      operationId: submitLearningAnswer
      tags: [learning]
      summary: Evaluate and record an active-learning answer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LearningAnswer"
      responses:
        "200":
          description: Evaluation feedback
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearningFeedback"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          description: AI service is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  parameters:
    CourseId:
      name: courseId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
    MaterialId:
      name: materialId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
    SessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    Course:
      type: object
      required:
        - id
        - name
        - description
        - materialCount
        - readyMaterialCount
        - progress
        - nextSession
        - activeLearning
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        materialCount:
          type: integer
        readyMaterialCount:
          type: integer
        progress:
          type: number
        nextSession:
          anyOf:
            - $ref: "#/components/schemas/StudySession"
            - type: "null"
        activeLearning:
          $ref: "#/components/schemas/ActiveLearningSettings"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CourseInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: string
          maxLength: 500
    CourseUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: string
          maxLength: 500
        activeLearning:
          $ref: "#/components/schemas/ActiveLearningSettings"
...
    MaterialInput:
      type: object
      required: [name, contentType, sizeBytes, objectPath]
      properties:
        name:
          type: string
          minLength: 1
        contentType:
          type: string
        sizeBytes:
          type: integer
          minimum: 1
        objectPath:
          type: string
          pattern: "^/objects/"
...
    DashboardSummary:
      type: object
      required: [courseCount, materialCount, readyMaterialCount, overallProgress, upcomingSessions, recentCourse]
      properties:
        courseCount:
          type: integer
        materialCount:
          type: integer
        readyMaterialCount:
          type: integer
        overallProgress:
          type: number
        upcomingSessions:
          type: array
          items:
            $ref: "#/components/schemas/StudySession"
        recentCourse:
          anyOf:
            - $ref: "#/components/schemas/Course"
            - type: "null"
...